% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_forecast_biomass.R
\name{app_forecast_biomass}
\alias{app_forecast_biomass}
\title{Forecast forest stocks for a stand}
\usage{
app_forecast_biomass()
}
\value{
This function runs a Shiny application
}
\description{
This function launches a Shiny application that allows users to
interactively run a forest stock forecast.
}
\details{
The app requires a CSV file in the "matrix model" format. To ensure
compliance, check the \emph{validate_wide_matrix()} function.
}
\examples{
\dontrun{
app_forecast_biomass()
}

}
