% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_select_stand.R
\name{app_select_stand}
\alias{app_select_stand}
\title{Select stands from an interactive map}
\usage{
app_select_stand()
}
\value{
This function runs a Shiny application
}
\description{
This function launches a Shiny application that allows users to interactively
select points from a map and manage selections in a table.
}
\details{
The app requires a CSV file with specific columns: Latitude, Longitude,
PrevB, PrevN, Hs, Hd, and PlotID. Users can select points individually or
using a rectangle selection tool.
}
\examples{
\dontrun{
app_select_stand()
}

}
