% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_model_cache.R
\name{clear_model_cache}
\alias{clear_model_cache}
\title{Clear model cache}
\usage{
clear_model_cache(model_name = NULL)
}
\arguments{
\item{model_name}{Specific model to remove from cache, or NULL to clear all}
}
\description{
Clear model cache
}
