% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_biomass.R
\name{forecast_biomass}
\alias{forecast_biomass}
\title{Project forest stocks over a plot file}
\usage{
forecast_biomass(
  save_to = NULL,
  data = NULL,
  plot_id = NULL,
  years = 50,
  clear_start = FALSE,
  clear_year = NULL,
  planting_init = NULL,
  planting_post = NULL,
  allow_colonization = TRUE,
  minimal_if_clearcut = TRUE,
  m_model = NULL,
  u_model = NULL,
  r_model = NULL,
  output_folder_name = "simulation_output"
)
}
\arguments{
\item{save_to}{Path to output folder}

\item{data}{A data frame or CSV file containing the input data in the
"matrix model" format.}

\item{plot_id}{Character. The identifier of the target plot.}

\item{years}{Numeric. Number of years to forecast.}

\item{clear_start}{Logical. Starts from clear cut. Defaults
to FALSE.}

\item{clear_year}{Numeric. A year to add a clearcut amid simulation.
Defaults to NULL (no clearcut).}

\item{planting_init}{A data frame with SPCD (species group) and TPH
(trees per hectare) columns. This is used to simulate growth of a new
plantation at the start if clear_start = TRUE. Default is NULL (no
plantation).}

\item{planting_post}{A data frame with SPCD (species group) and TPH
(trees per hectare) columns. This is used to simulate growth of a new
plantation after a mid-run clearcut (i.e., when clear_year is procided).
Default is NULL (no plantation).}

\item{allow_colonization}{Logical. Whether to allow other species to
colonize the stand. Defaults to FALSE (recruits only from the pre-cut
species pool.)}

\item{minimal_if_clearcut}{Logical.if any clear-cut happens, only save
summary and year_summary.}

\item{m_model}{Path to the mortality model '.rds' file.}

\item{u_model}{Path to the upgrowth model '.rds' file.}

\item{r_model}{Path to the recruitment model '.rds' file.}

\item{output_folder_name}{Character. Name for the output file.}
}
\value{
A list with data frames
}
\description{
Project forest stocks over a plot file
}
\examples{
\dontrun{
forecast_biomass(data = df, plot_id = "1234")
}
}
