% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_biomass_batch.R
\name{forecast_biomass_batch}
\alias{forecast_biomass_batch}
\title{Forecast forest stocks for many stands simultaneusly}
\usage{
forecast_biomass_batch(
  save_to = NULL,
  data_source = NULL,
  years = 50,
  clear_start = FALSE,
  clear_year = NULL,
  planting_init = NULL,
  planting_post = NULL,
  allow_colonization = TRUE,
  minimal_if_clearcut = TRUE,
  m_model = "./models/model_mortality.rds",
  u_model = "./models/model_upgrowth.rds",
  r_model = "./models/model_recruitment.rds",
  output_folder_name = "forecast_batch_results",
  n_cores = -1
)
}
\arguments{
\item{save_to}{Path to output folder}

\item{data_source}{A data frame, a list of data frames, a CSV file path,
or the path to a directory with multiple CSVs (individual plots).}

\item{years}{Numeric. Number of years to simulate.}

\item{clear_start}{Logical. Clear at first year. Defaults to FALSE.}

\item{clear_year}{Numeric. A year to add a clearcut amid simulation.
Defaults to NULL (no clearcut).}

\item{planting_init}{A data frame with SPCD (species group) and TPH
(trees per hectare) columns. This is used to simulate growth of a new
plantation at the start if clear_start = TRUE. Default is NULL (no
plantation).}

\item{planting_post}{A data frame with SPCD (species group) and TPH
(trees per hectare) columns. This is used to simulate growth of a new
plantation after a mid-run clearcut (i.e., when clear_year is procided).
Default is NULL (no plantation).}

\item{allow_colonization}{Logical. Whether to allow other species to
colonize the stand. Defaults to FALSE (recruits only from the pre-cut
species pool.)}

\item{minimal_if_clearcut}{Logical.if any clear-cut happens, only save
summary and year_summary.}

\item{m_model}{Path to the mortality model '.rds' file.}

\item{u_model}{Path to the upgrowth model '.rds' file.}

\item{r_model}{Path to the recruitment model '.rds' file.}

\item{output_folder_name}{Character. A name for the output folder to be
created in the 'save_to' directory.}

\item{n_cores}{Numeric. Number of cores to use in the parallel process.
Default is -1 (all but one).}
}
\value{
A list with data frames
}
\description{
Forecast forest stocks for many stands simultaneusly
}
\examples{
\dontrun{
forecast_biomass_batch(data = df_all)
}
}
