% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Get loaded model}
\usage{
get_model(model_name = c("mortality", "upgrowth", "recruitment"))
}
\arguments{
\item{model_name}{Name of model ("mortality", "upgrowth", or "recruitment")}
}
\value{
The loaded model
}
\description{
Get loaded model
}
