% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch a Shiny App from the package}
\usage{
launch_app(app_name = NULL)
}
\arguments{
\item{app_name}{The (partial or full) name of the Shiny app. If NULL,
lists available apps.}
}
\value{
Launches the Shiny app or lists available apps if no name is provided
}
\description{
Launch a Shiny App from the package
}
