% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{load_all_models}
\alias{load_all_models}
\title{Load all models from Zenodo}
\usage{
load_all_models(force_download = FALSE)
}
\arguments{
\item{force_download}{Whether to force re-download even if cached}
}
\description{
Load all models from Zenodo
}
