% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gif.R
\name{make_gif}
\alias{make_gif}
\title{Create GIF Animation from Images in Folder}
\usage{
make_gif(
  in_folder = NULL,
  out_folder = NULL,
  out_name = "animation",
  format = "png",
  fps = 2
)
}
\arguments{
\item{in_folder}{Character. Path to folder containing input images. Default is NULL.}

\item{out_folder}{Character. Path to folder for output GIF. Default is NULL.}

\item{out_name}{Character. A name for the output GIF file, without extension.}

\item{format}{Character. Image format to look for (e.g., "png", "jpg"). Default is "png".}

\item{fps}{Numeric. Frames per second for the animation. Default is 2.}
}
\value{
Creates a GIF file in the specified output folder. No return value.
}
\description{
Reads all images of specified format from a folder and creates an animated GIF.
}
\examples{
\dontrun{
make_gif(in_folder = "path/to/images",
         out_folder = "path/to/output",
         format = "png",
         fps = 5)
}

}
