% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixmodel-package.R
\docType{package}
\name{matrixmodel}
\alias{matrixmodel}
\alias{matrixmodel-package}
\title{matrixmodel: Matrix Model for Forest Biomass Projection}
\description{
The matrixmodel package provides functions for projecting forest biomass
using matrix models with random forest components, specifically designed
for Indiana forests in the US.
}
\section{Data}{

The package works with forest inventory data containing PlotID, SpeciesGroup,
Year, and biomass metrics.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/panchorivasf/matrixmodel}
  \item \url{https://panchorivasf.github.io/matrixmodel/}
  \item Report bugs at \url{https://github.com/panchorivasf/matrixmodel/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rivas \email{frivasfu@purdue.edu}

Authors:
\itemize{
  \item Jingjing Liang \email{jjliang@purdue.edu}
  \item Desmond Mensah \email{dmensah@purdue.edu}
  \item Varsha Raj Basavaraja \email{vbasavar@purdue.edu}
}

}
\keyword{internal}
