% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_means.R
\name{plot_means}
\alias{plot_means}
\title{Compute the plot-wise means for some variables in a "matrix model" data frame}
\usage{
plot_means(df)
}
\arguments{
\item{df}{A data frame in "matrix model" format}
}
\value{
A summary data frame with the means per plot
for selected variables
}
\description{
Compute the plot-wise means for some variables in a "matrix model" data frame
}
\examples{
\dontrun{
plot_values <- plot_means(df)
}
}
