% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pred_vector.R
\name{prepare_pred_vector}
\alias{prepare_pred_vector}
\title{Prepare the prediction vector}
\usage{
prepare_pred_vector(plt_vec, DBH)
}
\arguments{
\item{plt_vec}{The vector}

\item{DBH}{The DBH whatever}
}
\value{
A vector
}
\description{
Prepare the prediction vector
}
\examples{
prepare_pred_vector(plt_vec, DBH)

}
