% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_biomass.R
\name{project_biomass}
\alias{project_biomass}
\title{Project forest stocks over a plot file}
\usage{
project_biomass(
  save_to = NULL,
  data = NULL,
  plot_id = NULL,
  years = 50,
  m_model = "./models/model_mortality.rds",
  u_model = "./models/model_upgrowth.rds",
  r_model = "./models/model_recruitment1.rds",
  output_folder_name = "simulation_output"
)
}
\arguments{
\item{save_to}{Path to output folder}

\item{data}{A data frame or CSV file containing the input data in the matrix format.}

\item{plot_id}{Character. The identifier of the target plot.}

\item{years}{Numeric. Number of years to simulate.}

\item{m_model}{Path to the mortality model '.rds' file.}

\item{u_model}{Path to the upgrowth model '.rds' file.}

\item{r_model}{Path to the recruitment model '.rds' file.}

\item{output_folder_name}{Character. Name for the output file.}
}
\value{
A list with data frames
}
\description{
Project forest stocks over a plot file
}
\examples{
\dontrun{
project_biomass(data = df, plot_id = "1234")
}
}
