% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_biomass_batch.R
\name{project_biomass_batch}
\alias{project_biomass_batch}
\title{Project forest stocks for many plots}
\usage{
project_biomass_batch(
  save_to = NULL,
  data_source = NULL,
  years = 50,
  m_model = "./models/model_mortality.rds",
  u_model = "./models/model_upgrowth.rds",
  r_model = "./models/model_recruitment1.rds",
  n_cores = -1
)
}
\arguments{
\item{save_to}{Path to output folder}

\item{data_source}{A data frame, a list of data frames, a CSV file path,
or the path to a directory with multiple CSVs (individual plots).}

\item{years}{Numeric. Number of years to simulate.}

\item{m_model}{Path to the mortality model '.rds' file.}

\item{u_model}{Path to the upgrowth model '.rds' file.}

\item{r_model}{Path to the recruitment model '.rds' file.}
}
\value{
A list with data frames
}
\description{
Project forest stocks for many plots
}
\examples{
\dontrun{
project_biomass_batch(data = df_all)
}
}
