% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_model_options.R
\name{set_model_options}
\alias{set_model_options}
\title{Set package options for model loading}
\usage{
set_model_options(remote_base_url = NULL, cache_dir = NULL)
}
\arguments{
\item{remote_base_url}{Base URL for remote model repository}

\item{cache_dir}{Custom cache directory}
}
\description{
Set package options for model loading
}
