% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_diversity.R
\name{update_diversity}
\alias{update_diversity}
\title{Update diversity data}
\usage{
update_diversity(pred_vec)
}
\arguments{
\item{pred_vec}{Prediction vector}
}
\value{
An updated diversity data
}
\description{
Update diversity data
}
\examples{
\dontrun{
update_diversity()
}
}
