% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_predictions.R
\name{update_predictions}
\alias{update_predictions}
\title{Update predictions}
\usage{
update_predictions(pred_vec, m, u, r, DBH)
}
\arguments{
\item{pred_vec}{Predictions vector}

\item{m}{Mortality model}

\item{u}{Upgrowth model}

\item{r}{Recruitment model}

\item{DBH}{Diameter at breast height}
}
\value{
A prediction
}
\description{
Update predictions
}
\examples{
\dontrun{
update_preductions(pred_vec, m, u, r, DBH)
}
}
